Add sample1 project to demonstrate base design patterns
This project specifically shows the use of Salt with Vagrant and using
the defaults.yaml and map.jinja design patterns.  It also shows the use of
templating to create a set of states from pillar data and the subtle use
of the 'lookup' entry in a pillar vs types defined in the pillar.  It also
shows how to override variables for vagrant specific VMs using hostname as
a trigger.